fixing button for revealing rest of code
Initial proceed to checkout button wasn't supposed to include javascript with the html and was then removed. Button was also same as submit for bottom of form and was adjusted not to be "onclick" or a submit type. It was adjusted to include an id which would then be referenced later in the javascript below. Console logging would track the code to see if code executes correctly (revealed that we did not define the showhide so it will be adjusted to rebuild the function from scratch). A click function was added to allow javascript to execute the checkoutFunction.