Team images uploaded, jumbotron uploaded and applied, navbar fixed, padding applied to header and body, and some first elements, lorem ipsum adjusted for theme across all pages, card system implemented
Images for team were uploaded and organized in a card system, jumbotron added and applied to about page, padding was applied to header, body, and first element to avoid overlapping since nav was made to be fixed. Lorem ipsum for most areas adjusted from older content to fit a more hotel vibe.